+++ src/constants.py
@@ -1,0 +1,4 @@
+BASE_URL = 'https://stellarburgers.nomoreparties.site/'
+LOG_IN_URL = BASE_URL + '/login'
+REGISTRATION_URL = BASE_URL + '/register'
+FORGOT_PASSWORD_URL = BASE_URL + '/forgot-password'
