+++ tests/test_from_personal_account_to_constructor.py
@@ -1,6 +1,7 @@
 from selenium.webdriver.common.by import By
-from conftest import CreateUser, authoriztaion
+from helpers import CreateUser, authoriztaion
 from locators import LogInPage, LogInMainPageLocators, PersonalAccount, Constructor
+from src import constants
 
 
 class TestToConstructorFromPersonalAccount:
@@ -8,8 +9,7 @@ class TestToConstructorFromPersonalAccount:
         user = CreateUser.generate_user()
         email, password = user.email, user.password
         CreateUser(user.name, user.email, user.password).user_registration(driver, helpers)  # Регистрируем пользователя
-        url = 'https://stellarburgers.nomoreparties.site/'
-        driver.get(url)  # Переходим на главную
+        driver.get(constants.BASE_URL)  # Переходим на главную
 
         helpers.wait_until_visibility_of_element_located(driver, LogInMainPageLocators.LOG_IN_BUTTON)
         driver.find_element(By.XPATH, LogInMainPageLocators.LOG_IN_BUTTON).click()
@@ -27,5 +27,3 @@ class TestToConstructorFromPersonalAccount:
         driver.find_element(By.XPATH, LogInMainPageLocators.PERSONAL_ACCOUNT_LINK).click()
 
         driver.find_element(By.XPATH, Constructor.TO_CONSTRUCTOR_LOGO).click()
-
-        driver.quit()
